% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_table_of_contents.R
\name{build_and_use_gif}
\alias{build_and_use_gif}
\title{Title}
\usage{
build_and_use_gif(href, dir, pattern = "", which = 5, width = 100,
  height = 100, delay = 0.2, static_file = select_static(dir,
  pattern, which), gif_file = paste0(dir, "/viztoc_", pattern, ".gif"),
  num_in_gif = NULL, cached_gif = TRUE)
}
\arguments{
\item{delay}{}
}
\description{
Title
}
