% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_table_of_contents.R
\name{dir_pngs_build_static_gif_viztoc}
\alias{dir_pngs_build_static_gif_viztoc}
\title{Title}
\usage{
dir_pngs_build_static_gif_viztoc(dir, pattern = "", href, which = 5,
  width = 100, height = 100, delay = 0.2, pagedown = T,
  static_file = select_static(dir, pattern, which),
  gif_file = paste0(dir, "/viztoc_", pattern, ".gif"),
  num_in_gif = NULL, cached_gif = TRUE)
}
\arguments{
\item{delay}{}
}
\description{
Title
}
