% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_table_of_contents.R
\name{flipbook_webshots_build_and_use_gif}
\alias{flipbook_webshots_build_and_use_gif}
\title{Title}
\usage{
flipbook_webshots_build_and_use_gif(dir, base_url, href = base_url,
  num_pages, pattern = "", which = 5, width = 100, height = 100,
  delay = 0.2, static_file = select_static(dir, pattern, which),
  gif_file = paste0(dir, "/viztoc_", pattern, ".gif"),
  num_in_gif = NULL, cached_gif = TRUE)
}
\arguments{
\item{cached_gif}{}
}
\description{
Title
}
