% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_table_of_contents.R
\name{build_and_use_gif}
\alias{build_and_use_gif}
\title{Title}
\usage{
build_and_use_gif(href, dir, pattern, static_file = select_static(dir,
  pattern), gif_file = paste0(pattern, ".gif"), num_in_gif = NULL)
}
\arguments{
\item{num_in_gif}{}
}
\description{
Title
}
